requirements.txt
alembic==1.4.2
argcomplete>=1.11.1
factory_boy>=2.12.0
feedparser>=2.5.1
pytest>=5.4.1
pytest-cov
SQLAlchemy==1.3.15
